feat(case): unify R2 viewer at /case/ with Sketchfab fallback + dark default
- /case/?id=X probes R2; on 404, probes Sketchfab v3 API
  - Sketchfab 200 -> window.location.replace('/case/legacy/?id=X')
  - Sketchfab 404 -> friendly overlay asking the radiologist to upload
  - No ?id= at all -> "Nenhum caso foi informado neste link..."
- cache: "no-cache" on existence probes so a stale 404 cannot pin a
  freshly-uploaded case as missing
- Dark mode is now the default; "light" is honored when the user has
  explicitly toggled it (localStorage)
- Legacy Sketchfab viewer parked under /case/legacy/ (preserves measure,
  opacity, theme, video, laudo)
- /case-next/ kept as a redirect stub forwarding to /case/?<search>
- Landing-page CTAs now link to /case/ directly (no redirect hop)
- Tests: Sketchfab probe is mocked; new tests cover the
  R2-miss/Sketchfab-hit redirect and the both-miss overlay
- Test infra: dedicated http-server on :5505 with reuseExistingServer=false
  to stop colliding with other Conductor workspaces' static servers
