+++ case/legacy/index.html
@@ -1,0 +1,66 @@
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <link rel="preconnect" href="https://fonts.googleapis.com">
+  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+  <link href="https://fonts.googleapis.com/css2?family=Nunito+Sans:wght@400;600;700&family=Open+Sans:wght@400;600&display=swap" rel="stylesheet">
+  <link rel="stylesheet" href="../../style.css" />
+  <script type="text/javascript" src="https://static.sketchfab.com/api/sketchfab-viewer-1.9.0.js"></script>
+</head>
+
+<body style="background-color: #111;">
+  <!-- VIDEO button directly in body, not inside any div -->
+  <button id="global_trigger" data-is-hidden="false"
+    style="display: none; position: fixed; top: 16px; left: 16px; margin-left: 0px !important; color: white; background-color: rgb(59, 57, 57); border: 1px solid white; z-index: 100; padding: 0px 0px; border-radius: 8px; cursor: pointer; font-size: 16px;">
+    Vídeo
+  </button>
+  <div class="min-height-90vh" style="background-color: #272425 !important; z-index: 10000;">
+    <div class="py-1 pb-3 container">
+
+
+
+      <!-- Botão Medir - aligned with Estruturas panel -->
+      <button class="btn-outline-primary" id="newMeasure">
+        <img id="imggMedir" src="measure-icon.svg" width="24" height="24" alt="Measure" />
+        Medir
+      </button>
+
+
+      <div class="row" style="padding-top: 70px;">
+
+        <div class="d-flex flex-column col-md-8 skfb-view">
+          <div class="top-cover" id="top"></div>
+          <iframe id="api-frame" width="100%" height="70vh"></iframe>
+
+          <div class="logo-cover" id="logo"></div>
+          <!-- Botão Mudar Fundo no canto inferior esquerdo -->
+          <button id="bg-black"
+            style="position: absolute; bottom: 0px;margin-left: 0px !important; left: 0px; background: #272425; z-index: 10000;">
+            <img id="icon" src="moon.svg" alt="modo escuro" width="24">
+          </button>
+          <div class="iframe-overlay-top" id="overtop"></div>
+          <div class="iframe-overlay-bottom" id="overbottom"></div>
+        </div>
+
+
+        <!-- Estruturas mais para baixo -->
+        <div id="structures" class="d-flex flex-column col-md-4 mx-auto rounded"
+          style="color: white; background-color:rgb(59, 57, 57); margin-top: 20px; z-index: 999;">
+          <div id="title" style="color: white; background-color: rgb(59, 57, 57); max-width: 600px">
+            <label id="sub" class="Subtitle"
+              style="color: white; background-color:  rgb(59, 57, 57);">Estruturas</label>
+          </div>
+          <div class="navTree" id="navTree"></div>
+        </div>
+      </div>
+    </div>
+  </div>
+
+  <script src="main.js"></script>
+  <script src="measure.js"></script>
+  <script src="opacity.js"></script>
+  <script src="laudo.js"></script>
+  <script src="mudaCor.js"></script>
+
+  <script src="botao_video.js"></script>
+</body>
