+++ case/legacy/opacity.js
