+++ index.html
@@ -139,7 +139,7 @@
             <p class="lede">Reconstruções 3D de alta fidelidade entregues em até três dias. Visualização interativa para cirurgia complexa — no consultório, na sala, no celular do plantonista.</p>
             <div class="cta-row">
               <a href="https://wa.me/5521993118288?text=Ol%C3%A1%2C%20quero%20saber%20mais%20sobre%20as%20solu%C3%A7%C3%B5es%20em%20modelos%203D%20para%20cirurgias." class="btn-primary coral" target="_blank" rel="noopener">Solicitar um caso →</a>
-              <a href="https://biodesignlab.com.br/case-next/?id=5fc6c4d2d77d4ab6a8cadfe8996c70a4" target="_blank" rel="noopener" class="btn-ghost">
+              <a href="https://biodesignlab.com.br/case/?id=5fc6c4d2d77d4ab6a8cadfe8996c70a4" target="_blank" rel="noopener" class="btn-ghost">
                 <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="1.8" style="width:16px;height:16px;"><polygon points="5 3 19 12 5 21 5 3"/></svg>
                 Abrir caso demo
               </a>
@@ -329,7 +329,7 @@
             <p class="lede">Envie o exame na unidade Dasa. Em até três dias úteis o link do modelo 3D chega no seu celular, pronto para girar, medir e compartilhar com a equipe.</p>
             <div class="cta-row" style="margin-top:var(--space-6);">
               <a href="https://wa.me/5521993118288?text=Ol%C3%A1%2C%20quero%20saber%20mais%20sobre%20as%20solu%C3%A7%C3%B5es%20em%20modelos%203D%20para%20cirurgias." class="btn-primary coral" target="_blank" rel="noopener">Solicitar um caso →</a>
-              <a href="https://biodesignlab.com.br/case-next/?id=5fc6c4d2d77d4ab6a8cadfe8996c70a4" target="_blank" rel="noopener" class="btn-ghost-dark">
+              <a href="https://biodesignlab.com.br/case/?id=5fc6c4d2d77d4ab6a8cadfe8996c70a4" target="_blank" rel="noopener" class="btn-ghost-dark">
                 <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="1.8" style="width:16px;height:16px;"><polygon points="5 3 19 12 5 21 5 3"/></svg>
                 Abrir caso demo
               </a>
@@ -353,7 +353,7 @@
           <a href="/" class="wm">Medcase<em>.</em></a>
           <p style="font-size:13px;color:#A8B0B6;max-width:32ch;margin:0;">Visualizador 3D cirúrgico.</p>
         </div>
-        <div><h4>Produto</h4><ul><li><a href="/case-next/">Visualizador</a></li><li><a href="/upload/">Upload de caso</a></li></ul></div>
+        <div><h4>Produto</h4><ul><li><a href="/case/">Visualizador</a></li><li><a href="/upload/">Upload de caso</a></li></ul></div>
         <div><h4>Equipe</h4><ul><li><a href="#equipe">Médicos responsáveis</a></li></ul></div>
         <div><h4>Concierge</h4><ul><li><a href="tel:+5521993118288">+55 21 99311-8288</a></li><li><a href="https://wa.me/5521993118288?text=Ol%C3%A1%2C%20quero%20saber%20mais%20sobre%20as%20solu%C3%A7%C3%B5es%20em%20modelos%203D%20para%20cirurgias." target="_blank" rel="noopener">WhatsApp</a></li><li><a href="mailto:contato@biodesignlab.com.br">contato@biodesignlab.com.br</a></li></ul></div>
       </div>
