+++ playwright.config.js
@@ -7,14 +7,14 @@ export default defineConfig({
   reporter: "list",
   timeout: 15_000,
   use: {
-    baseURL: "http://127.0.0.1:5500",
+    baseURL: "http://127.0.0.1:5505",
     trace: "retain-on-failure",
     screenshot: "only-on-failure",
   },
   webServer: {
-    command: "npx http-server -p 5500 -c-1 --silent",
-    url: "http://127.0.0.1:5500",
-    reuseExistingServer: true,
+    command: "npx http-server -p 5505 -c-1 --silent",
+    url: "http://127.0.0.1:5505",
+    reuseExistingServer: false,
     timeout: 10_000,
   },
   projects: [
